python code
# 🎲 Number Guessing Game (Python) By DIVYANSH PANDEY

A simple **Python console game** where the computer randomly selects a number between **1 and 100**, and the player has to guess it.  
The program gives hints like **"Too High"** or **"Too Low"** until the player guesses the correct number.  
It also counts the number of attempts the player takes to find the answer.

---

## 🚀 Features
- Random number generated every time you run the game 🎲  
- Hints provided (`Too Low` / `Too High`) 🔼🔽  
- Counts how many attempts you took 🧮  
- Simple, beginner-friendly Python project 🐍  

---

## 📂 How to Run
 Clone this repository:
   ```run it on any online or offline interpreter